refactor(library): Organise abstraction layers into nested subfolders
