+++ src/lib/CRSFforArduino/src/SerialReceiver/CRSF/CRSF.h
@@ -28,7 +28,7 @@
 
 #include "Arduino.h"
 #include "CRSFProtocol.h"
-#include "CRC.h"
+#include "SerialReceiver/CRC/CRC.h"
 // #include "Hardware.h"
 
 namespace serialReceiver
